Updated migration check existence colukmn and table

+++ alembic/versions/0014_add_is_deleted.py
@@ -22,10 +22,14 @@ from sqlalchemy.engine.reflection import Inspector
 def upgrade() -> None:
     conn = op.get_bind()
     inspector = Inspector.from_engine(conn)
-    columns = [col['name'] for col in inspector.get_columns('post_prod_projects')]
-    if 'is_deleted' not in columns:
-        op.add_column('post_prod_projects', sa.Column('is_deleted', sa.Boolean(), server_default='false', nullable=False))
+    if inspector.has_table('post_prod_projects'):
+        columns = [col['name'] for col in inspector.get_columns('post_prod_projects')]
+        if 'is_deleted' not in columns:
+            op.add_column('post_prod_projects', sa.Column('is_deleted', sa.Boolean(), server_default='false', nullable=False))
 
 
 def downgrade() -> None:
-    op.drop_column('post_prod_projects', 'is_deleted')
+    conn = op.get_bind()
+    inspector = Inspector.from_engine(conn)
+    if inspector.has_table('post_prod_projects'):
+        op.drop_column('post_prod_projects', 'is_deleted')
